Split CommentUIIntegrationTest to extensions
