+++ EssentialApp/EssentialAppTests/CommentUIIntegrationTests.swift
@@ -165,73 +165,4 @@ class CommentUIIntegrationTests: XCTestCase {
 		let presentableModel = makePresentableComment(comment: model)
 		return (model, presentableModel)
 	}
-	
-	private func assertThat(_ sut: CommentViewController, isRendering comments: [PresentableComment], file: StaticString = #file, line: UInt = #line) {
-		
-		guard sut.numberOfRenderedComments() == comments.count else {
-			return XCTFail("Expected \(comments.count) comments, got \(sut.numberOfRenderedComments()) instead", file: file, line: line)
-		}
-		
-		comments.enumerated().forEach { index, comment in
-			assertThat(sut, hasViewConfiguredFor: comment, at: index, file: file, line: line)
-		}
-	}
-	
-	private func assertThat(_ sut: CommentViewController, hasViewConfiguredFor comment: PresentableComment, at index: Int, file: StaticString = #file, line: UInt = #line) {
-		let view = sut.commentView(at: index)
-		guard let cell = view as? CommentCell else {
-			return XCTFail("Expected to get \(CommentCell.self), got \(String(describing: view)) instead")
-		}
-
-		XCTAssertEqual(cell.authorText, comment.author, "Expected `authorText` to be \(String(describing: cell.authorText)) for cell at index \(index)", file: file, line: line)
-		XCTAssertEqual(cell.messageText, comment.message, "Expected `messageText` to be \(String(describing: cell.messageText)) for cell at index \(index)", file: file, line: line)
-		XCTAssertEqual(cell.timestampText, comment.createAt, "Expected `timestampText` to be \(String(describing: cell.timestampText)) for cell at index \(index)", file: file, line: line)
-	}
 }
-
-private extension CommentViewController {
-	func simulateUserInititateCommentReload() {
-		refreshControl?.simulatePullToRefresh()
-	}
-	
-	var isShowingLoadingIndicator: Bool {
-		return refreshControl?.isRefreshing == true
-	}
-	
-	func numberOfRenderedComments() -> Int {
-		return tableView.numberOfRows(inSection: commentSection)
-	}
-	
-	func commentView(at row: Int) -> UITableViewCell? {
-		let ds = tableView.dataSource
-		let index = IndexPath(row: row, section: commentSection)
-		return ds?.tableView(tableView, cellForRowAt: index)
-	}
-	
-	func simulateViewDisappearance() {
-		self.viewWillDisappear(true)
-	}
-	
-	var isShowingErrorView: Bool {
-		return errorView?.message != nil
-	}
-	
-	private var commentSection: Int {
-		return 0
-	}
-}
-
-extension CommentCell {
-	var authorText: String? {
-		return authorLabel?.text
-	}
-	
-	var messageText: String? {
-		return commentLabel?.text
-	}
-	
-	var timestampText: String? {
-		return timestampLabel?.text
-	}
-}
-
