+++ EssentialApp/EssentialAppTests/Helpers/FeedViewController+TestHelpers.swift
@@ -68,4 +68,10 @@ extension FeedViewController {
 	private var feedImagesSection: Int {
 		return 0
 	}
+	
+	func simulateUserSelectFeed(at row: Int) {
+		let delegate = tableView.delegate
+		let index = IndexPath(row: row, section: feedImagesSection)
+		delegate?.tableView?(tableView, didSelectRowAt: index)
+	}
 }
